Added deleting items from inventory (back end and front, tentatively, with a basic button) and searching for items on the back end. Both back-end functions return JSON responses.
For deleting items from the inventory on the front-end, POST to {{ url_for('deleteitem').replace(request.script_root,'') }} (which should be '/deleteitem') with the ID number of the item as variable "id". It will return a JSON object (call it 'obj', why not) with obj['success'] = true or false.

For searching from the front-end, POST to {{ url_for('receiveSearchForm').replace(request.script_root,'') }} (should resolve to '/browse') with, as variables sent, "query": the search string, "price": an integer, and "category": the underlying text values of any of the radio buttons in the search form. It will return a JSON with obj['items'] = <list of items i with attributes i['name'], i['category'],
i['price'], i['imgurl'] (URL for the image), i['description'], i['ID'] (capitalized, it's the database URL of the item, which is actually not very important and shouldn't show on the page), and i['user'] (the user ID, same as that in the database table 'users'). I'll replace that last one with the poster user's name instead of ID, but just use the number for the time being when displaying it.

This should be done with an AJAX request, maybe using jQuery or just good ol' vanilla JS, and should dynamically populate the HTML page.